--bug=1022346 --user=李明辉 【授权应用】授权应用选择可调用次数为无限次，点击确定，接口报错 --bug=1022354 --user=李明辉 【应用授权】后端需更改编辑授权的调用次数后的计算逻辑，现是每编辑一次重新计算，PRD描述为累计计算，详见描述

+++ icreditstudio-dataapi/src/main/java/com/jinninghui/datasphere/icreditstudio/dataapi/enums/CallCountEnum.java
@@ -7,11 +7,11 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum CallCountEnum {
 
-    CALL_FINITE_TIMES(0L, 0),
-    CALL_INFINITE_TIMES(-1L, 1)
+    CALL_FINITE_TIMES(0, 0),
+    CALL_INFINITE_TIMES(-1, 1)
     ;
 
-    private Long code;
+    private Integer code;
     private Integer callTime;
 
 }
